Fix debug message not logged using ocf_log
The code was expecting HA_debug to be != 0 when debug was turn on in Pacemaker.
When HA_debug was not set, the OCF_functions module was actually setting this env
var to 0 to disable it, leading to this bug.
It turns out this HA_debug env var is not set anymore under Centos 7 (and probably
in most other distrib I guess). The only clue about this var appears in code related
to the AIS stack. Tests like  "x$HA_debug" = "x0" just work in shell (debug was
considered active).

Now, OCF_functions does not force a value to the HA_debug env var anymore and
make proper existance test when needed.